+++ packages/react-magma-docs/src/themes/amgam.ts
@@ -13,6 +13,6 @@ export const amgam = Object.assign({}, magma, {
     success: '#C57DFF',
     success02: '#AC44FF',
     danger: '#18FFFF',
-    focus: '#0599FF'
-  })
+    focus: '#0599FF',
+  }),
 });