+++ packages/react-magma-dom-codesandbox-template/src/App.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import './styles.css';
+import { GlobalStyles } from 'react-magma-dom';
+import { Example } from './Example';
+
+export default function App() {
+  return (
+    <div className="App">
+      <GlobalStyles />
+      <Example />
+    </div>
+  );
+}