+++ packages/react-magma-dom-codesandbox-template/src/Example.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Alert, AlertVariant } from 'react-magma-dom';
+export function Example() {
+  return (
+    <div>
+      <Alert variant={AlertVariant.danger}>Hello</Alert>
+    </div>
+  );
+}