+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.3.4-alpha.0 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
+### 2.3.3 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
+### 2.3.3-alpha.10 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
 ### 2.3.3-alpha.9 (2021-01-19)
 
 **Note:** Version bump only for package react-magma-dom