+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.0.8-alpha.0 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.10 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
 ### 1.0.7-alpha.9 (2021-01-19)
 
 **Note:** Version bump only for package react-magma-icons