+++ packages/react-magma-landing/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.0.3-alpha.0 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-landing
+
+
+
+
+
+### 1.0.2 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-landing
+
+
+
+
+
+### 1.0.2-alpha.10 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-landing
+
+
+
+
+
 ### 1.0.2-alpha.9 (2021-01-19)
 
 **Note:** Version bump only for package react-magma-landing